# Conflicts:
#	src/main/java/com/rivalhub/event/match/ResultsController.java

+++ src/main/java/com/rivalhub/common/FileUploadUtil.java
@@ -45,11 +45,13 @@ public class FileUploadUtil {
         return uploadDir + "/" + fileName;
     }
 
-    public void updateUserImage(UserData requestUser, MultipartFile multipartFile) {
-        if (multipartFile == null) {
+    public void updateUserImage(UserData requestUser, MultipartFile multipartFile, boolean keepAvatar) {
+
+        if (multipartFile == null && !keepAvatar) {
             deleteFileIfExists(requestUser);
             return;
-        }
+        } else if (multipartFile==null) return;
+
         String fileName = createFileName(multipartFile, userImgName);
         String uploadDir = getUploadDir(requestUser, multipartFile);
